Moves internals out of ui folder
Resolves https://github.com/brave/brave-browser/issues/4003
