+++ browser/resources/resource_ids
@@ -70,6 +70,10 @@
   "<(ROOT_GEN_DIR)/brave/web-ui-brave_donate/brave_donate.grd": {
     "includes": [38800],
   },
+  # This file is generated during the build.
+  "<(ROOT_GEN_DIR)/brave/web-ui-brave_rewards_internals/brave_rewards_internals.grd": {
+    "includes": [38900],
+  },
   "brave/components/resources/brave_components_strings.grd": {
     "messages": [39000]
   },
