+++ components/brave_rewards/resources/BUILD.gn
@@ -22,6 +22,15 @@ pack_web_resources("donate_generated_resources") {
   ]
 }
 
+pack_web_resources("internals_generated_resources") {
+  resource_name = "brave_rewards_internals"
+  output_dir = "$root_gen_dir/brave/components/brave_rewards/resources"
+  # need webpack and GRD generation first
+  deps = [
+    "internals"
+  ]
+}
+
 grit("static_resources") {
   source = "brave_rewards_static_resources.grd"
 
@@ -47,12 +56,14 @@ repack("resources") {
   deps = [
     ":ui_generated_resources",
     ":donate_generated_resources",
+    ":internals_generated_resources",
     ":static_resources",
   ]
 
   sources = [
     "$root_gen_dir/brave/components/brave_rewards/resources/brave_rewards_generated.pak",
     "$root_gen_dir/brave/components/brave_rewards/resources/brave_donate_generated.pak",
+    "$root_gen_dir/brave/components/brave_rewards/resources/brave_rewards_internals_generated.pak",
     "$root_gen_dir/brave/components/brave_rewards/resources/brave_rewards_static.pak",
   ]
