+++ components/brave_rewards/resources/brave_rewards_static_resources.grd
@@ -22,7 +22,7 @@
       <include name="IDR_BRAVE_REWARDS_FAVICON" file="img/favicon.ico" type="BINDATA" />
 
       <!-- WebUI Brave Rewards Internals resources -->
-      <include name="IDR_BRAVE_REWARDS_INTERNALS_HTML" file="ui/internals/brave_rewards_internals.html" type="BINDATA" />
+      <include name="IDR_BRAVE_REWARDS_INTERNALS_HTML" file="internals/brave_rewards_internals.html" type="BINDATA" />
 
       <!-- WebUI donate resources -->
       <include name="IDR_BRAVE_DONATE_BG_1" file="img/donate_bg_1.svg" type="BINDATA" />
