+++ components/brave_rewards/resources/internals/reducers/index.ts
