+++ components/brave_rewards/resources/internals/reducers/rewards_internals_reducer.ts
