+++ components/brave_rewards/resources/internals/storage.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { debounce } from '../../../../common/debounce'
+import { debounce } from '../../../common/debounce'
 
 const keyName = 'rewards-internals-data'
